requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
pillow
realesrgan
basicsr
gfpgan